refactor: 로컬스토리지 토큰 관리 및 리다이렉션 로직 리팩토링
