+++ src/apis/apiCall.ts
@@ -1,5 +1,6 @@
 import { NetworkError, UnauthorizedError } from "@/errors/errors";
 import type { ApiResponse } from "./common/types";
+import { ACCESS_TOKEN, REFRESH_TOKEN } from "@/const/localStorage";
 
 export const API_BASE = import.meta.env.VITE_API_BASE_URL;
 
@@ -13,8 +14,8 @@ const tokenRefresh = async (refreshToken: string) => {
   })
     .then((res) => res.json())
     .then((res) => {
-      localStorage.setItem("access-token", res.data.accessToken);
-      localStorage.setItem("refresh-token", res.data.refreshToken);
+      localStorage.setItem(ACCESS_TOKEN, res.data.accessToken);
+      localStorage.setItem(REFRESH_TOKEN, res.data.refreshToken);
     })
     .catch((e) => {
       throw e;
@@ -30,7 +31,7 @@ export const apiCall = async <T>(
 ): Promise<ApiResponse<T>> => {
   let SAFE_RETRY_FLAG = 0;
 
-  const access_token = localStorage.getItem("access-token");
+  const access_token = localStorage.getItem(ACCESS_TOKEN);
   if (withAuth && !access_token) {
     throw new UnauthorizedError();
   }
@@ -66,8 +67,8 @@ export const apiCall = async <T>(
     })
     .then((res) => res.json())
     .catch((e) => {
-      localStorage.removeItem("access-token");
-      localStorage.removeItem("refresh-token");
+      localStorage.removeItem(ACCESS_TOKEN);
+      localStorage.removeItem(REFRESH_TOKEN);
 
       if (e instanceof TypeError) {
         throw new NetworkError();
