+++ src/apis/auth/authAPI.ts
@@ -1,3 +1,4 @@
+import { ACCESS_TOKEN, REFRESH_TOKEN } from "@/const/localStorage";
 import { apiCall } from "../apiCall";
 import type { ApiResponse } from "../common/types";
 import type { AuthResponse, logoutResponse } from "./authTypes";
@@ -6,8 +7,8 @@ import { getResponseType } from "./authUtils";
 export const sendAuthCode = async (authCode: string, addr: string) => {
   const data = { code: authCode, redirectUri: addr };
   const res: ApiResponse<AuthResponse> = await apiCall("/oauth/google", "POST", data, false);
-  localStorage.setItem("access-token", res.data.accessToken);
-  localStorage.setItem("refresh-token", res.data.refreshToken);
+  localStorage.setItem(ACCESS_TOKEN, res.data.accessToken);
+  localStorage.setItem(REFRESH_TOKEN, res.data.refreshToken);
   return res;
 };
 
@@ -16,8 +17,8 @@ export const logout = async () => {
   const type = getResponseType(res.code);
 
   if (type === "success") {
-    sessionStorage.removeItem("access-token");
-    sessionStorage.removeItem("refresh-token");
+    localStorage.removeItem(ACCESS_TOKEN);
+    localStorage.removeItem(REFRESH_TOKEN);
     return res;
   } else {
     throw Promise.reject(new Error(res.message));
