+++ src/const/localStorage.ts
@@ -1,0 +1,2 @@
+export const ACCESS_TOKEN = "access-token";
+export const REFRESH_TOKEN = "refresh-token";
