+++ src/pages/login/components/LoginContainer.tsx
@@ -1,18 +1,10 @@
-import { useEffect } from "react";
 import Meetcha_banner from "../../../components/Meetcha_banner";
 import Continue_Google from "./Continue_Google";
 import Continue_des from "./Continue_des";
 
 import "../styles/login.scss";
-import { useNavigate } from "react-router-dom";
 
 const LoginContainer = () => {
-  const navigate = useNavigate();
-  useEffect(() => {
-    if (localStorage.getItem("access-token")) {
-      navigate("/schedule");
-    }
-  }, []);
   return (
     <div className="login_container">
       <div className="flex_container1">
